atualizei o codigo des 31

+++ Aulas Curso Em vídeo/aula010/des 31.py
@@ -1,5 +1,5 @@
 print("Qual a distância da sua viagem?")
-dis= int(input("Distância (em km): "))
+dis= float(input("Distância (em km): "))
 if dis <= 200:
     val= 0.5 * dis
     print("a sua passagem custará {}".format(val))
